Catch IFTTT exception
An exception from this call is happening a large number of times a day. By
catching it we should drastically cut down on cluter in our logs and also
avoid any retrying that might be going on due to this.

Note that in sumo logic we are reporting mostly org.xml.sax.SAXParseException
however that is because we are running "sanitizeRootCause" on the exception.
If you inspect the full stacktrace you will see that this is the top level
error thrown, and locally this does catch bad xml being returned as expected.

Note I don't know how to test this so I mainly tested the syntax in a
different app, but have not been able to test this in any environment.
If anybody has a good way to test this I am open to ideas, but I want
to get something out there to avoid throwing this exception all the time.